Make generic List parametrized
